requirements.txt
google-api-python-client==1.7.4
google-cloud-firestore==0.30.0
oauth2client==4.1.3